+++ features/Context/EnterpriseNavigationContext.php
@@ -13,9 +13,10 @@ use Context\NavigationContext as BaseNavigationContext;
 class EnterpriseNavigationContext extends BaseNavigationContext
 {
     protected $enterprisePageMapping = [
-        'published' => 'Published index',
-        'proposals' => 'Proposal index',
-        'assets' => 'Asset index'
+        'published'      => 'Published index',
+        'proposals'      => 'Proposal index',
+        'assets'         => 'ProductAsset index',
+        'asset edit'     => 'ProductAsset edit'
     ];
 
     /**
@@ -32,6 +33,28 @@ class EnterpriseNavigationContext extends BaseNavigationContext
     public function iShouldBeOnThePublishedProductsPage()
     {
         $expectedAddress = $this->getPage('Published index')->getUrl();
+        $this->assertAddress($expectedAddress);
+    }
+
+    /**
+     * @param string $asset
+     *
+     * @Given /^I should be on the "([^"]+)" asset edit page$/
+     */
+    public function iShouldBeOnTheAssetEditPage($assetCode)
+    {
+        $asset = $this->getMainContext()->getSubcontext('fixtures')->getProductAsset($assetCode);
+
+        $expectedAddress = $this->getPage('ProductAsset edit')->getUrl(['id' => $asset->getId()]);
+        $this->assertAddress($expectedAddress);
+    }
+
+    /**
+     * @Given /^I should be on the assets page$/
+     */
+    public function iShouldBeOnTheAssetsPage()
+    {
+        $expectedAddress = $this->getPage('ProductAsset index')->getUrl();
         $this->assertAddress($expectedAddress);
     }
 }